gatos

+++ Seminario 1/src/persistencia/PlanDAO.java
@@ -27,7 +27,7 @@ public class PlanDAO {
 
 	public static boolean existePlanVigente() {
 		Query q  = s.createQuery("select count(p) from Plan p where p.estado = 3 "); 
-		int cant  = (int)q.uniqueResult();
+		long cant  = (long)q.uniqueResult();
 		if(cant > 0 )
 			return true;
 		else
